Revert "Merge branch 'main' of https://github.com/microsoft/pyright into perf/benchmark-suite"
This reverts commit abaa715ec3281294835b710d71bbe53a21204e91, reversing
changes made to d8dd0a212e88d6283b7aece092a319aa7f9a2a16.
