+++ .github/workflows/mypy_primer_comment.yaml
@@ -18,7 +18,7 @@ jobs:
   comment:
     name: Comment PR from mypy_primer
     runs-on: ubuntu-latest
-    if: ${{ (github.repository == 'microsoft/pyright' || github.repository == 'bschnurr/pyright') && github.event.workflow_run.conclusion == 'success' }}
+    if: ${{ github.event.workflow_run.conclusion == 'success' }}
     steps:
       - name: Download diffs
         uses: actions/github-script@v7
