+++ .github/workflows/mypy_primer_pr.yaml
@@ -29,7 +29,6 @@ concurrency:
 
 jobs:
   mypy_primer:
-    if: github.repository == 'microsoft/pyright' || github.repository == 'bschnurr/pyright'
     name: Run mypy_primer on PR
     runs-on: ubuntu-latest
     permissions:
