+++ .github/workflows/mypy_primer_push.yaml
@@ -22,7 +22,6 @@ concurrency:
 
 jobs:
   mypy_primer:
-    if: github.repository == 'microsoft/pyright' || github.repository == 'bschnurr/pyright'
     name: Run mypy_primer on push
     runs-on: ubuntu-latest
     permissions:
