--- .github/workflows/pyright_ecosystem_benchmark.yaml
@@ -1,130 +1,0 @@
-name: Pyright ecosystem benchmark
-
-on:
-  workflow_dispatch:
-    inputs:
-      mode:
-        description: 'Run mode'
-        required: true
-        default: compare
-        type: choice
-        options:
-          - compare
-          - refresh-baseline
-      project_date:
-        description: 'Project checkout date passed to the ecosystem runner'
-        required: true
-        default: '2026-01-01'
-        type: string
-      project:
-        description: 'Optional project name regex filter'
-        required: false
-        default: ''
-        type: string
-      install_dependencies:
-        description: 'Install synced ecosystem project dependencies before running Pyright'
-        required: false
-        default: false
-        type: boolean
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}-${{ inputs.mode }}
-  cancel-in-progress: true
-
-jobs:
-  ecosystem-smoke:
-    if: github.repository == 'microsoft/pyright' || github.repository == 'bschnurr/pyright'
-    name: Ecosystem smoke benchmark
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: '20'
-
-      - uses: actions/setup-python@v5
-        with:
-          python-version: '3.11'
-
-      - name: Install dependencies
-        run: npm ci
-
-      - name: Build Pyright CLI and benchmark runner
-        run: |
-          npm run build:cli:dev
-          cd packages/pyright-internal
-          npm run build
-          npm run bench:ecosystem:sync
-
-      - name: Run ecosystem smoke comparison
-        if: ${{ inputs.mode == 'compare' }}
-        shell: bash
-        run: |
-          set -euo pipefail
-          cd packages/pyright-internal
-
-          baseline_path="./src/tests/benchmarks/baselines/ecosystem-smoke-main.json"
-          if [[ ! -f "$baseline_path" ]]; then
-            echo "Missing checked-in ecosystem smoke baseline at $baseline_path" >&2
-            exit 1
-          fi
-
-          args=(
-            --suite smoke
-            --project-root "$GITHUB_WORKSPACE/.ecosystem-projects"
-            --prepare-projects
-            --project-date "${{ inputs.project_date }}"
-            --candidate-executable "node ../pyright/index.js"
-            --main-baseline-report "$baseline_path"
-            --output ./src/tests/benchmarks/.generated/benchmark-results/ecosystem-pr
-          )
-
-          if [[ -n "${{ inputs.project }}" ]]; then
-            args+=(--project "${{ inputs.project }}")
-          fi
-
-          if [[ "${{ inputs.install_dependencies }}" == "true" ]]; then
-            args+=(--install-dependencies)
-          fi
-
-          npm run bench:ecosystem:run -- "${args[@]}"
-
-      - name: Refresh ecosystem smoke baseline
-        if: ${{ inputs.mode == 'refresh-baseline' }}
-        shell: bash
-        run: |
-          set -euo pipefail
-          cd packages/pyright-internal
-
-          args=(
-            --suite smoke
-            --project-root "$GITHUB_WORKSPACE/.ecosystem-projects"
-            --prepare-projects
-            --project-date "${{ inputs.project_date }}"
-            --output ./src/tests/benchmarks/.generated/benchmark-results/ecosystem-main
-            --baseline-source-commit "$GITHUB_SHA"
-          )
-
-          if [[ -n "${{ inputs.project }}" ]]; then
-            args+=(--project "${{ inputs.project }}")
-          fi
-
-          if [[ "${{ inputs.install_dependencies }}" == "true" ]]; then
-            args+=(--install-dependencies)
-          fi
-
-          npm run bench:ecosystem:update-main-baseline -- "${args[@]}"
-
-      - name: Upload ecosystem benchmark artifacts
-        uses: actions/upload-artifact@v4
-        with:
-          name: pyright-ecosystem-benchmark-${{ inputs.mode }}
-          path: |
-            packages/pyright-internal/src/tests/benchmarks/.generated/benchmark-results/
-            packages/pyright-internal/src/tests/benchmarks/baselines/ecosystem-smoke-main.json
-          if-no-files-found: warn
