+++ packages/pyright-internal/package.json
@@ -16,11 +16,6 @@
         "test": "npm run webpack:testserver && node --max-old-space-size=8192 --expose-gc ./node_modules/jest/bin/jest --forceExit --testPathIgnorePatterns src/tests/benchmarks",
         "test:norebuild": "node --max-old-space-size=8192 --expose-gc ./node_modules/jest/bin/jest --forceExit --testPathIgnorePatterns src/tests/benchmarks",
         "test:benchmark": "cross-env PYRIGHT_RUN_BENCHMARKS=1 node --max-old-space-size=8192 --expose-gc ./node_modules/jest/bin/jest --forceExit --testTimeout=300000 --runInBand --detectOpenHandles src/tests/benchmarks",
-        "bench:ecosystem:run": "node ./out/packages/pyright-internal/src/tests/benchmarks/runEcosystemBenchmark.js",
-        "bench:ecosystem:run:local": "node ./out/packages/pyright-internal/src/tests/benchmarks/runEcosystemBenchmark.js --baseline-executable \"node ../pyright/index.js\" --candidate-executable \"node ../pyright/index.js\"",
-        "bench:ecosystem:update-main-baseline": "node ./out/packages/pyright-internal/src/tests/benchmarks/runEcosystemBenchmark.js --baseline-executable \"node ../pyright/index.js\" --update-main-baseline",
-        "bench:ecosystem:sync": "node ./out/packages/pyright-internal/src/tests/benchmarks/syncMypyPrimerProjects.js",
-        "bench:ecosystem:smoke": "node ./out/packages/pyright-internal/src/tests/benchmarks/runEcosystemBenchmark.js --suite smoke --output ./src/tests/benchmarks/.generated/benchmark-results/ecosystem-run",
         "test:coverage": "node --max-old-space-size=8192 --expose-gc ./node_modules/jest/bin/jest --forceExit --testPathIgnorePatterns src/tests/benchmarks --reporters=jest-junit --reporters=default --coverage --coverageReporters=cobertura --coverageReporters=html --coverageReporters=json",
         "test:imports": "node --max-old-space-size=8192 --expose-gc ./node_modules/jest/bin/jest importResolver.test --forceExit --runInBand"
     },
