+++ packages/pyright-internal/src/tests/benchmarks/tokenizerBenchmark.test.ts
@@ -13,21 +13,20 @@
  *   src/tests/benchmarks/.generated/benchmark-results/tokenizer/
  */
 
+import { execFileSync } from 'child_process';
+import * as fs from 'fs';
+import * as os from 'os';
+import * as path from 'path';
+
 import { Tokenizer } from '../../parser/tokenizer';
-import {
-    calculateStats,
-    createBenchmarkReport,
-    formatCount,
-    loadBenchmarkCorpus,
-    runJestBenchmarkInFreshProcess,
-    writeBenchmarkReport,
-} from './benchmarkUtils';
 
 // --- Configuration ---
 
 const WARMUP_ITERATIONS = 3;
 const BENCHMARK_ITERATIONS = 10;
 
+const BENCHMARK_OUTPUT_DIR = path.join(__dirname, '.generated', 'benchmark-results', 'tokenizer');
+const JEST_BIN_PATH = path.resolve(__dirname, '..', '..', '..', 'node_modules', 'jest', 'bin', 'jest.js');
 const CHILD_RESULT_PREFIX = '__TOKENIZER_BENCHMARK_RESULT__';
 const CHILD_MODE_ENV = 'PYRIGHT_TOKENIZER_BENCH_CHILD';
 const RUN_BENCHMARKS_ENV = 'PYRIGHT_RUN_BENCHMARKS';
@@ -48,7 +47,69 @@ interface BenchmarkResult {
     tokensPerSec: number;
 }
 
+interface BenchmarkReport {
+    timestamp: string;
+    system: {
+        platform: string;
+        arch: string;
+        cpus: string;
+        cpuCount: number;
+        totalMemoryMB: number;
+        nodeVersion: string;
+    };
+    config: {
+        warmupIterations: number;
+        benchmarkIterations: number;
+    };
+    results: BenchmarkResult[];
+}
+
 // --- Helpers ---
+
+function calculateStats(times: ReadonlyArray<number>): {
+    median: number;
+    p95: number;
+    min: number;
+    max: number;
+    avg: number;
+} {
+    const sorted = [...times].sort((a, b) => a - b);
+    const len = sorted.length;
+
+    const median = len % 2 === 0 ? (sorted[len / 2 - 1] + sorted[len / 2]) / 2 : sorted[Math.floor(len / 2)];
+    const p95Index = Math.ceil(len * 0.95) - 1;
+    const p95 = sorted[Math.min(p95Index, len - 1)];
+    const min = sorted[0];
+    const max = sorted[len - 1];
+    const avg = times.reduce((a, b) => a + b, 0) / len;
+
+    return { median, p95, min, max, avg };
+}
+
+function loadCorpus(filename: string): string {
+    const filePath = path.resolve(__dirname, '..', 'benchmarkData', filename);
+    return fs.readFileSync(filePath, 'utf-8');
+}
+
+function getSystemInfo(): BenchmarkReport['system'] {
+    const cpus = os.cpus();
+    return {
+        platform: os.platform(),
+        arch: os.arch(),
+        cpus: cpus[0]?.model ?? 'unknown',
+        cpuCount: cpus.length,
+        totalMemoryMB: Math.round(os.totalmem() / (1024 * 1024)),
+        nodeVersion: process.version,
+    };
+}
+
+function writeReport(report: BenchmarkReport): void {
+    fs.mkdirSync(BENCHMARK_OUTPUT_DIR, { recursive: true });
+    const filename = `tokenizer-benchmark-${new Date().toISOString().replace(/[:.]/g, '-')}.json`;
+    const outputPath = path.join(BENCHMARK_OUTPUT_DIR, filename);
+    fs.writeFileSync(outputPath, JSON.stringify(report, undefined, 2), 'utf-8');
+    console.log(`\nBenchmark results written to: ${outputPath}`);
+}
 
 function printResultTable(results: ReadonlyArray<BenchmarkResult>): void {
     console.log('\n=== Tokenizer Benchmark Results ===\n');
@@ -68,7 +129,7 @@ function printResultTable(results: ReadonlyArray<BenchmarkResult>): void {
                 .toFixed(2)
                 .padStart(10)} ${result.avgMs.toFixed(2).padStart(10)} ${result.p95Ms
                 .toFixed(2)
-                .padStart(10)} ${formatCount(result.tokensPerSec).padStart(12)}`
+                .padStart(10)} ${Math.round(result.tokensPerSec).toLocaleString().padStart(12)}`
         );
     }
     console.log('');
@@ -78,14 +139,55 @@ function emitChildResult(result: BenchmarkResult): void {
     process.stdout.write(`${CHILD_RESULT_PREFIX}${JSON.stringify(result)}\n`);
 }
 
+function getChildOutput(error: unknown): string {
+    if (!(error instanceof Error)) {
+        return '';
+    }
+
+    const stdout = 'stdout' in error && typeof error.stdout === 'string' ? error.stdout : '';
+    const stderr = 'stderr' in error && typeof error.stderr === 'string' ? error.stderr : '';
+    return [stdout, stderr].filter((part) => part.length > 0).join('\n');
+}
+
+function escapeRegExp(text: string): string {
+    return text.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+}
+
 function runBenchmarkInFreshProcess(testName: string): BenchmarkResult {
-    return runJestBenchmarkInFreshProcess(
-        __filename,
-        'Tokenizer Benchmark',
-        testName,
-        CHILD_RESULT_PREFIX,
-        CHILD_MODE_ENV
-    );
+    try {
+        const output = execFileSync(
+            process.execPath,
+            [
+                JEST_BIN_PATH,
+                __filename,
+                '--runInBand',
+                '--forceExit',
+                '--testTimeout=300000',
+                '--testNamePattern',
+                `^Tokenizer Benchmark ${escapeRegExp(testName)}$`,
+            ],
+            {
+                cwd: path.resolve(__dirname, '..', '..', '..'),
+                encoding: 'utf-8',
+                env: {
+                    ...process.env,
+                    [CHILD_MODE_ENV]: '1',
+                },
+            }
+        );
+
+        const resultLine = output.split(/\r?\n/).find((line) => line.startsWith(CHILD_RESULT_PREFIX));
+
+        if (!resultLine) {
+            throw new Error(`Child benchmark for "${testName}" did not emit a result.\n${output}`);
+        }
+
+        return JSON.parse(resultLine.slice(CHILD_RESULT_PREFIX.length)) as BenchmarkResult;
+    } catch (error) {
+        const output = getChildOutput(error);
+        const message = error instanceof Error ? error.message : String(error);
+        throw new Error(`Child benchmark for "${testName}" failed.\n${message}${output ? `\n${output}` : ''}`);
+    }
 }
 
 function benchmarkTokenize(corpusName: string, code: string): BenchmarkResult {
@@ -148,7 +250,7 @@ benchmarkSuite('Tokenizer Benchmark', () => {
     for (const { name, file } of corpora) {
         test(`tokenize ${name}`, () => {
             const result = isChildProcess
-                ? benchmarkTokenize(name, loadBenchmarkCorpus(file))
+                ? benchmarkTokenize(name, loadCorpus(file))
                 : runBenchmarkInFreshProcess(`tokenize ${name}`);
 
             if (!isChildProcess) {
@@ -156,9 +258,9 @@ benchmarkSuite('Tokenizer Benchmark', () => {
             }
 
             console.log(
-                `  ${name}: median=${result.medianMs.toFixed(2)}ms, tokens=${result.tokenCount}, tok/sec=${formatCount(
+                `  ${name}: median=${result.medianMs.toFixed(2)}ms, tokens=${result.tokenCount}, tok/sec=${Math.round(
                     result.tokensPerSec
-                )}`
+                ).toLocaleString()}`
             );
 
             if (isChildProcess) {
@@ -172,7 +274,7 @@ benchmarkSuite('Tokenizer Benchmark', () => {
 
     test('scaled corpus (10x large_stdlib)', () => {
         const result = isChildProcess
-            ? benchmarkTokenize('large_stdlib_10x', Array(10).fill(loadBenchmarkCorpus('large_stdlib.py')).join('\n'))
+            ? benchmarkTokenize('large_stdlib_10x', Array(10).fill(loadCorpus('large_stdlib.py')).join('\n'))
             : runBenchmarkInFreshProcess('scaled corpus (10x large_stdlib)');
 
         if (!isChildProcess) {
@@ -182,7 +284,7 @@ benchmarkSuite('Tokenizer Benchmark', () => {
         console.log(
             `  large_stdlib_10x: median=${result.medianMs.toFixed(2)}ms, tokens=${
                 result.tokenCount
-            }, tok/sec=${formatCount(result.tokensPerSec)}`
+            }, tok/sec=${Math.round(result.tokensPerSec).toLocaleString()}`
         );
 
         if (isChildProcess) {
@@ -199,10 +301,16 @@ benchmarkSuite('Tokenizer Benchmark', () => {
 
         printResultTable(allResults);
 
-        writeBenchmarkReport(
-            'tokenizer',
-            'tokenizer-benchmark',
-            createBenchmarkReport('tokenizer', WARMUP_ITERATIONS, BENCHMARK_ITERATIONS, allResults)
-        );
+        const report: BenchmarkReport = {
+            timestamp: new Date().toISOString(),
+            system: getSystemInfo(),
+            config: {
+                warmupIterations: WARMUP_ITERATIONS,
+                benchmarkIterations: BENCHMARK_ITERATIONS,
+            },
+            results: allResults,
+        };
+
+        writeReport(report);
     });
 });
